FixDbSpace2: delete folder class, add file DbSpaceInfo
