+++ com.cubrid.cubridmanager.core/src/com/cubrid/cubridmanager/core/cubrid/dbspace/model/DbSpaceInfoList.java
@@ -41,7 +41,6 @@ import com.cubrid.common.core.util.LogUtil;
 import com.cubrid.cubridmanager.core.common.model.EnvInfo;
 import com.cubrid.cubridmanager.core.common.model.IModel;
 import com.cubrid.cubridmanager.core.common.model.ServerVersion;
-import com.cubrid.cubridmanager.core.common.task.CommonQueryTask;
 
 /**
  *
@@ -51,6 +50,7 @@ import com.cubrid.cubridmanager.core.common.task.CommonQueryTask;
  * @author sq
  * @version 1.0 - 2009-12-28 created by sq
  */
+
 public class DbSpaceInfoList implements
 		IModel {
 	
@@ -65,142 +65,13 @@ public class DbSpaceInfoList implements
 		}
 	}
 	
-	public static class DbSpaceInfo {
-
-		private String spacename = null;
-
-		/*
-		 * GENERIC,DATA,TEMP,INDEX,Active_log,Archive_log
-		 */
-		private String type = null;
-		private String location = null;
-		private int totalpage = 0;
-		private int freepage = 0;
-		private int usedpage = 0;
-		private int volid;
-		private String date = null;
-		private String purpose;
-		private int volumeCount = 0;
-		private String totalPageStr = null;
-		private String totalSizeStr = null;
-
-		public String getTotalPageStr() {
-			return totalPageStr;
-		}
-
-		public void setTotalPageStr(String totalPageStr) {
-			this.totalPageStr = totalPageStr;
-		}
-
-		public String getTotalSizeStr() {
-			return totalSizeStr;
-		}
-
-		public void setTotalSizeStr(String totalSizeStr) {
-			this.totalSizeStr = totalSizeStr;
-		}
-
-		public String getSpacename() {
-			return spacename;
-		}
-
-		public void setSpacename(String spacename) {
-			this.spacename = spacename;
-		}
-
-		public String getType() {
-			return type;
-		}
-
-		public void setType(String type) {
-			this.type = type;
-		}
-
-		public String getLocation() {
-			return location;
-		}
-
-		public void setLocation(String location) {
-			this.location = location;
-		}
-
-		public int getTotalpage() {
-			return totalpage;
-		}
-
-		public void setTotalpage(int totalpage) {
-			this.totalpage = totalpage;
-		}
-
-		public int getFreepage() {
-			return freepage;
-		}
-
-		public void setFreepage(int freepage) {
-			this.freepage = freepage;
-		}
-		
-		public int getUsedpage() {
-			return usedpage;
-		}
-		
-		public void setUsedpage(int usedpage){
-			this.usedpage = usedpage;
-		}
-
-		public String getDate() {
-			return date;
-		}
-
-		public void setDate(String date) {
-			this.date = date;
-		}
-		
-		public int getVolid() {
-			return volid;
-		}
-		
-		public void setVolid(int volid) {
-			this.volid = volid;
-		}
-
-		public int getVolumeCount() {
-			return volumeCount;
-		}
-		
-		public String getPurpose() {
-			return purpose;
-		}
-		
-		public void setPurpose(String purpose) {
-			this.purpose = purpose;
-		}
-		
-		public String getShortVolumeName(){
-			int lastIndex;
-			if ((lastIndex = spacename.lastIndexOf('/')) >= 0){
-				return spacename.substring(lastIndex+1);
-			} else {
-				return spacename;
-			}
-		}
-
-		/**
-		 * Get the next volume count value
-		 */
-		public void plusVolumeCount() {
-			this.volumeCount++;
-		}
-
-	}
-	
 	private static final Logger LOGGER = LogUtil.getLogger(DbSpaceInfoList.class);
 	protected String dbname = null;
 	protected int pagesize = 0;
 	protected int logpagesize = 0;
 	protected int freespace = 0;
 	
-	protected List<DbSpaceInfoList.DbSpaceInfo> spaceinfo = null;
+	protected List<DbSpaceInfo> spaceinfo = null;
 	private static final ServerVersion changedFormatVersion = new ServerVersion(10, 1);
 	
 	/***
@@ -209,7 +80,7 @@ public class DbSpaceInfoList implements
 	 * @return List<DbSpaceInfo> the list that encapsulates the instances of
 	 *         DbSpaceInfo
 	 */
-	public List<DbSpaceInfoList.DbSpaceInfo> getSpaceinfo() {
+	public List<DbSpaceInfo> getSpaceinfo() {
 		synchronized (this) {
 			return spaceinfo;
 		}
@@ -221,7 +92,7 @@ public class DbSpaceInfoList implements
 	 * @return Map<String, DbSpaceInfo> the map that encapsulates the instances
 	 *         of DbSpaceInfo
 	 */
-	public Map<String, DbSpaceInfoList.DbSpaceInfo> getSpaceInfoMap() {
+	public Map<String, DbSpaceInfo> getSpaceInfoMap() {
 		synchronized (this) {
 			Map<String, DbSpaceInfo> map = new TreeMap<String, DbSpaceInfo>();
 			if (spaceinfo == null) {
@@ -261,7 +132,7 @@ public class DbSpaceInfoList implements
 	 * @param spaceinfoList List<DbSpaceInfo> A list that encapsulates the
 	 *        instances of DbSpaceInfo
 	 */
-	public void setSpaceinfo(List<DbSpaceInfoList.DbSpaceInfo> spaceinfoList) {
+	public void setSpaceinfo(List<DbSpaceInfo> spaceinfoList) {
 		synchronized (this) {
 			this.spaceinfo = spaceinfoList;
 		}
@@ -273,7 +144,7 @@ public class DbSpaceInfoList implements
 	 *
 	 * @param info DbSpaceInfo A instance of DbSpaceInfo
 	 */
-	public void addSpaceinfo(DbSpaceInfoList.DbSpaceInfo info) {
+	public void addSpaceinfo(DbSpaceInfo info) {
 		synchronized (this) {
 			if (spaceinfo == null) {
 				spaceinfo = new ArrayList<DbSpaceInfo>();
@@ -290,7 +161,7 @@ public class DbSpaceInfoList implements
 	 *
 	 * @param info DbSpaceInfo A instance of DbSpaceInfo
 	 */
-	public void removeSpaceinfo(DbSpaceInfoList.DbSpaceInfo info) {
+	public void removeSpaceinfo(DbSpaceInfo info) {
 		synchronized (this) {
 			if (spaceinfo != null) {
 				spaceinfo.remove(info);
